represent TCP ports as `uint16`s
TCP ports are [defined as 16-bit unsigned integers][1]. Currently,
however, the Go controllers often represent them as other numeric types,
such as `int` and `uint32`. We currently validate parsed ports by testing
whether they are less than 65536. However, it's not always clear whether
a given `int` represents a port number that has already been validated
by this check or not.

This commit changes the `PortRange` type and parsing code to represent
ports as `uint16`, making invalid states harder to represent at the type
level. We change port parsing from using `strconv.Atoi` (which returns
an `int`, which may be negative or greater than 65535) to using
`strconv.ParseUint` with a bit size of 16 --- this function will fail to
parse the value if it is a negative number or greater than 65535, so we
no longer have to perform additional validation when parsing ports. We
can also remove some casting that is no longer necessary. It's still
necessary to cast to `uint32` in a couple of places when constructing
protobuf messages, because protobuf doesn't support 16-bit integers.

Depends on #10827

[1]: https://en.wikipedia.org/wiki/Port_(computer_networking)#Port_number

+++ controller/api/destination/endpoint_translator.go
@@ -32,7 +32,7 @@ type endpointTranslator struct {
 	enableH2Upgrade     bool
 	nodeTopologyZone    string
 	nodeName            string
-	defaultOpaquePorts  map[uint32]struct{}
+	defaultOpaquePorts  map[uint16]struct{}
 
 	availableEndpoints watcher.AddressSet
 	filteredSnapshot   watcher.AddressSet
@@ -48,7 +48,7 @@ func newEndpointTranslator(
 	enableH2Upgrade bool,
 	service string,
 	srcNodeName string,
-	defaultOpaquePorts map[uint32]struct{},
+	defaultOpaquePorts map[uint16]struct{},
 	k8sAPI *k8s.MetadataAPI,
 	stream pb.Destination_GetServer,
 	log *logging.Entry,
@@ -253,7 +253,7 @@ func (et *endpointTranslator) sendClientAdd(set watcher.AddressSet) {
 	for _, address := range set.Addresses {
 		var (
 			wa          *pb.WeightedAddr
-			opaquePorts map[uint32]struct{}
+			opaquePorts map[uint16]struct{}
 			err         error
 		)
 		if address.Pod != nil {
@@ -348,7 +348,7 @@ func toAddr(address watcher.Address) (*net.TcpAddress, error) {
 	}, nil
 }
 
-func createWeightedAddr(address watcher.Address, opaquePorts map[uint32]struct{}, enableH2Upgrade bool, identityTrustDomain string, controllerNS string, log *logging.Entry) (*pb.WeightedAddr, error) {
+func createWeightedAddr(address watcher.Address, opaquePorts map[uint16]struct{}, enableH2Upgrade bool, identityTrustDomain string, controllerNS string, log *logging.Entry) (*pb.WeightedAddr, error) {
 
 	tcpAddr, err := toAddr(address)
 	if err != nil {
@@ -375,7 +375,7 @@ func createWeightedAddr(address watcher.Address, opaquePorts map[uint32]struct{}
 	controllerNSLabel := address.Pod.Labels[pkgK8s.ControllerNSLabel]
 	sa, ns := pkgK8s.GetServiceAccountAndNS(address.Pod)
 	weightedAddr.MetricLabels = pkgK8s.GetPodLabels(address.OwnerKind, address.OwnerName, address.Pod)
-	_, isSkippedInboundPort := skippedInboundPorts[address.Port]
+	_, isSkippedInboundPort := skippedInboundPorts[uint16(address.Port)]
 
 	// If the pod is controlled by any Linkerd control plane, then it can be
 	// hinted that this destination knows H2 (and handles our orig-proto
@@ -386,7 +386,7 @@ func createWeightedAddr(address watcher.Address, opaquePorts map[uint32]struct{}
 		// opaque by annotation or default value, then hint its proxy's
 		// inbound port, and set the hinted protocol to Opaque, so that the
 		// client proxy does not send a SessionProtocol.
-		_, opaquePort := opaquePorts[address.Port]
+		_, opaquePort := opaquePorts[uint16(address.Port)]
 		if address.OpaqueProtocol || opaquePort {
 			port, err := getInboundPort(&address.Pod.Spec)
 			if err != nil {
